Update issue templates
CABOOTLE is a METAVERSE ambassador to a new frontier of social interactivity and commerce. The universal scope and collaboration is literally taking the power of the planet to manifest this new future. With that being said the old adages "each one teach one" and "it takes a village" are necessary in order to educate to average person of the benefits these tremendous technological advances. Just like with any other leap in technology society had to have an adjustment period before change was readily accepted. This is a similar time to when email was getting explained to the public and it totally killed the long-distance industry. CABOOTLE will he here to assist with the evolution of this brand new technology.